mosaico-packs/pyproject.toml
[tool.poetry]
name = "mosaico-packs"
version = "0.1.0"
description = "Packs for Mosaico Data-Platform"
authors = ["Mosaico Labs <info@mosaico.dev>"]
readme = "README.md"
license = "AGPL-3.0"
keywords = [
    "mosaicolabs",
    "mosaico-sdk",
    "mosaico-packs",
    "data-platform",
    "physical-ai",
    "robotics",
]

packages = [{ include = "mosaicopacks", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<4"
mosaicolabs = { path = "../mosaico-sdk-py", develop = true }
numpy = ">=2.0.0,<2.5.0"
pillow = ">=11.3.0,<13.0.0"
rich = "^14.3.3"
rosbags = "^0.11.0"
av = "^17.0.0"
h5py = ">=3.10.0"
tensorflow = ">=2.15.0"
tensorflow-datasets = ">=4.9.0"
protobuf = ">=3.20.0,<4.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.4.2"
ruff = "^0.15.8"
pre-commit = "^4.5.1"

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["COM812", "E501"]

[tool.ruff.format]
skip-magic-trailing-comma = false
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.isort]
combine-as-imports = true

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
"mosaicopacks" = "mosaicopacks.cli:run_pack_cli"
